% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcInitialPCA.R
\name{calcInitialPCA}
\alias{calcInitialPCA}
\title{Initial PCA of  attributes from sample files originatting from Illumina GenomeStudio}
\usage{
calcInitialPCA(ht12object)
}
\arguments{
\item{ht12object}{A list object of class HT12prepro}
}
\value{
a list object including slots `pca_object` which is a copy of the $chipsamples  with additional columns representing the principal components. Additionally, `pc.data` is the PCA-object created by function princomp()
}
\description{
Numerical Attributes from the slot `$chipsamples` from an HT12prepro object are dimension-reduced by PCR
}
