% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkExtractChipsamples.R
\name{checkExtractChipsamples}
\alias{checkExtractChipsamples}
\title{Check Illumina HT12v4 expression data}
\usage{
checkExtractChipsamples(paramfile = NULL,
  file_names_of_files_and_folders = "from_paramfile",
  colseparator = "from_paramfile", renameDublettes = "from_paramfile",
  prefix_new_samplename = "from_paramfile")
}
\arguments{
\item{paramfile}{Path to the file specifying preprocessing parameters, can be NULL}

\item{file_names_of_files_and_folders}{Path to a tab-delimeted file with information of files created in Illuminas Genome Studio.  Following columns are required:   `fileset_id`: a short ID for a fileset,  `con_f`: full path and filename of the control probe data file,  `nobkgd_f`: full path and filename of the gene probe data file,  `sample_f`: full path and filename of the control probe data file. If NULL, information is used from parameterfile specified in 'paramfile', Default: 'from_paramfile'}

\item{colseparator}{Separator splitting numbers into units of 1000. Typically "." or  ",". Default: 'from_paramfile'}

\item{renameDublettes}{If any sample Ids is found more than once in the sample file provided by Illumina, suffixes .1 .2 etc are added to the newly created new_ID used to designate samples within the preprocessing. If FALSE, the function will stop when duplicated Sample Ids are found. Default: 'from_paramfile'}

\item{prefix_new_samplename}{Prefix of new_ID, the new R-compatibel sample names used to designate samples within the preprocessing. Default: 'from_paramfile'}
}
\value{
a list of class `HT12prepro` including a slot with sample-related attributes of the current processing-stage named `$chipsamples`
}
\description{
This function checks the data and extracts the sample-IDs including information from Illuminas sample-files. See vignette for an example.
}
