% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareEsetExpressions.R
\name{compareEsetExpressions}
\alias{compareEsetExpressions}
\title{compareEsetExpressions}
\usage{
compareEsetExpressions(all_nobkgd_eset_ql_combat, adjusted_eset,
  firstmessage = NULL, showPlots = T, compmethod = "spearman")
}
\arguments{
\item{all_nobkgd_eset_ql_combat}{PARAM_DESCRIPTION}

\item{adjusted_eset}{PARAM_DESCRIPTION}

\item{firstmessage}{PARAM_DESCRIPTION, Default: NULL}

\item{showPlots}{PARAM_DESCRIPTION, Default: T}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
