% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparePCBeforeAfterPrePro.R
\name{comparePCBeforeAfterPrePro}
\alias{comparePCBeforeAfterPrePro}
\title{Visualize PCA before and after preprocessing expression levels}
\usage{
comparePCBeforeAfterPrePro(ht12object, paramfile = NULL, showPlots = T)
}
\arguments{
\item{ht12object}{A list object of class HT12prepro created with function visualizePreprocessing}

\item{paramfile}{Path to the file specifying parameters}
}
\value{
A list object of class HT12prepro including updated slots, e.g   the slot with the history of the commands named `$history`` is updated. Plots are also stored in slot `$`
}
\description{
PCA on expression data is done using good samples surviving preprocessing. Following PCAs are calculated and shown as figures: PCA of all transcripts before preprocessing, including additional scaling, PCA of gene expression after preprocessing using only expression probes that are expressed, not overinflated regarding batch affects and beeing specific to the human genome when considering a remapping approach according to Barbosa-Moralis et al. (2010)
}
