% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createExpressionSet.R
\name{createExpressionSet}
\alias{createExpressionSet}
\title{Create an expressionset from Illumina HT12v4 data}
\usage{
createExpressionSet(ht12object, paramfile = NULL,
  colseparator = "from_paramfile")
}
\arguments{
\item{ht12object}{A list object of class HT12prepro created with function checkExtractChipsamples()}

\item{paramfile}{Path to the file specifying parameters}

\item{colseparator}{Separator splitting numbers into units of 1000. Typically "." or  ",". If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}
}
\value{
A list object of class HT12prepro with a slot with updated  sample-related attributes of the current processing-stage named `$chipsamples`, a slot with the R-object holding expression data used to create the expression set named `$rawdataWOcons_joined`, a slot with the R-object holding control data used to create the expression set named `$rawdataOnlycons_joined`, a slot with the R-object holding control data used to create the expression set named `$rawdataOnlycons_joined`, a slot with expression-set expression data excluding control data named `$all_nobkgd_eset`, a slot with expression-set expression data including control data named `$total_nobkgd_eset`, a slot with the history of the commands named `$history`including expressionset slots
}
\description{
This function loads the expression data from several text files created by Illumina GenomeStudio. These files are specified via the slot $chipsamples of an HT12prepro-object from columns "fileset_id", "nobkgd_f", "con_f",and "sample_f". See vignette for an example.
}
