% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTextForMethods.R
\name{createTextForMethods}
\alias{createTextForMethods}
\title{Create Text for Material and Methods to report preprocessing}
\usage{
createTextForMethods(ht12object, paramfile)
}
\arguments{
\item{ht12object}{A list object of class HT12prepro}

\item{paramfile}{Path to the file specifying parameters}
}
\value{
an object with a markdown formated short and long text describing preprocessing
}
\description{
Creates Text for Material and Methods to report preprocessing
}
