% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter4MinBatchsize.R
\name{filter4MinBatchsize}
\alias{filter4MinBatchsize}
\title{Filter samples of a HT12prepro object for min batch size n>1}
\usage{
filter4MinBatchsize(
  ht12object,
  paramfile = NULL,
  second_combat_withvar = "from_paramfile"
)
}
\arguments{
\item{ht12object}{A list object of class HT12prepro created with function filterTechnicallyFailed()}

\item{paramfile}{Path to the file specifying parameters}

\item{second_combat_withvar}{A second round of removeBatchEffects() can be intended. In this case, batches from the first round checking Sentrix.Barcode batches as well as batches from this second round specified as a column in ht12object$chipsamples named according to this variable must be checked for minimum batch size n>1.If "", no 2nd combat will be done. If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}
}
\value{
A list object of class HT12prepro where the  slot with  sample-related attributes of the current processing-stage named `$chipsamples` is update. Excluded individual are characterized by column in_study ==F and reason4exclusion
}
\description{
This functions filters for min Batch size n>1  as this is a limitation for following removal of batch effects in the following step applying function removeBatchEffects()
}
