% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAtypicalExpressed.R
\name{filterAtypicalExpressed}
\alias{filterAtypicalExpressed}
\title{Filter samples with atypical expressoin data}
\usage{
filterAtypicalExpressed(
  ht12object,
  paramfile = NULL,
  filter2ind_atypischEuklid = "from_paramfile"
)
}
\arguments{
\item{ht12object}{A list object of class HT12prepro created with function checkBatchEffects()}

\item{paramfile}{Path to the file specifying parameters}

\item{filter2ind_atypischEuklid}{Filter for extreme combination of values from expressed transcripts summarized as Euclidian distance of expression values of all expression probes previously classified as expressed and non-overinflated regarding ANOVA on batch-effects.  Valid is ln('Euclidian distance') < median(ln('Euclidian Distance' )) + [value] *  IQR(ln('Euclidian Distance' )). If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}
}
\value{
A list object of class HT12prepro where the slot with  sample-related attributes of the current processing-stage named `$chipsamples` is updated.  Excluded individual are characterized by column in_study ==F and reason4exclusion. The slot with the history of the commands named `$history`` is updated.
}
\description{
Filter samples for atypical gene expression levels. This is defined as atypical large Euclidian distance of all expressed, not batch-associated expression probes to an artificial individual. This artificial individual was defined as the average of samples after removing 10% samples farthest away from the average of all samples done separately for each subgroup (implemented in the R / Bioconductor package lumi [PMID:18467348]).
}
