% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLowExpressed.R
\name{filterLowExpressed}
\alias{filterLowExpressed}
\title{Filter samples with atypical number of expressed genes and annotate probes for their expression level}
\usage{
filterLowExpressed(ht12object, paramfile = NULL,
  filter1ind_expressedGenes = "from_paramfile",
  filter1probes_expressedProbes = "from_paramfile")
}
\arguments{
\item{ht12object}{A list object of class HT12prepro created with function createExpressionSet()}

\item{paramfile}{Path to the file specifying parameters.}

\item{filter1ind_expressedGenes}{Filter for extreme number of 'Detected.Genes..0.01.' at detection p-value 0.01. Valid is: 'Detected.Genes..0.01.' < Median - [value]x IQR AND > Median - [value]x IQR.  If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile}

\item{filter1probes_expressedProbes}{Attribute asigning an expression probe as 'expressed' if detected within at least [value]x n(individuals) at detection p-value 0.05.  If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}
}
\value{
A list object of class HT12prepro  where the slot with  sample-related attributes of the current processing-stage named `$chipsamples` is updated.  Excluded individual are characterized by column in_study ==F and reason4exclusion. Additionally, a slot with the detailed probe-related expression level information named `$genesdetail is created, and the slot with the history of the commands named `$history`` is updated.
}
\description{
This function filters from an HT12prepro object samples with atypical number of expressed genes and annotates probes for their respective expression level
}
