% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterTechnicallyFailed.R
\name{filterTechnicallyFailed}
\alias{filterTechnicallyFailed}
\title{Identify and filter samples with unusual combination of control probes in a HT12prepro object}
\usage{
filterTechnicallyFailed(ht12object, paramfile = NULL,
  filter2ind_atypischIlmnKontroll = "from_paramfile",
  controlparameters2check = "from_paramfile",
  robustmethod_for_mahal = "from_paramfile")
}
\arguments{
\item{ht12object}{A list object of class HT12prepro created with function transformNormalizeHT12object()}

\item{paramfile}{Path to the file specifying parameters}

\item{filter2ind_atypischIlmnKontroll}{Filter for extreme combination of Illuminas control features summarized as Mahalanobis distance to an artificial sample. This artificial sample having has average values for the selected quality control features. Valid is ln('Mahalanobis distance') < median(ln('Mahalanobis distance')) + [value] *  IQR(ln('Mahalanobis distance')). If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}

\item{controlparameters2check}{Illuminas control features used for calculation of the Mahalanobis distance based outlyer criterium. Default ist 'hybrid_low, hybrid_med,  string_pm, string_mm, biotin, negative' (if spiked in erccc is used in ALL CHIPS use also 'ercc_1, ercc_2, ercc_3, ercc_4, ercc_5', if spked in artificial polyadenylated RNAs from Bacillus subtilis immediately preceding the reverse transcription step is used, also 'labeling'). In case of population based studies within a single tissue default is also 'housekeeping, Detected.Genes..0.01.'.  If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}

\item{robustmethod_for_mahal}{Method used for calculation of robust correllation for Mahalanobis distance within function mdqc(). Previous default was "S-estimator" (which has a 25 percent breakdown point), now "MVE" is prefferred (i.e. the Minimum Volume Ellipsoid (MVE) which searches for the ellipsoid with the smallest volume that covers h data points). Alternative is "MCD" (i.e. the Minimum Covariance Determinant (MCD) estimator which looks for the subset of h data points whose covariance matrix has the smallest determinant). If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}
}
\value{
A list object of class HT12prepro including updated slot `$chipsamples` with  sample-related attributes of the current processing-stage. Excluded individual are characterized by column in_study ==F and reason4exclusion. QC plots are shown.
}
\description{
log2 transformed and normalized  not heavily correlated control probes of the still valid samples are used to create a correlation-adjusted score (via the Mahalanobis distance of these parameters) reflecting technical accuracy of the expression measurement. Outlyers are identified defined as having a greater distance than `filter2ind_atypischIlmnKontroll` interquantile ranges. See vignette for an example.
}
