% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh.R
\name{hh}
\alias{hh}
\title{Show first rows and columns}
\usage{
hh(d, mydims = 5)
}
\arguments{
\item{d}{object to display}

\item{mydims}{how many rows of how many columns to show}
}
\value{
Matrix of the first mydims rows of the first mydims columns
}
\description{
Displays the first mydims rows of the first mydims columns
}
