% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ht.R
\name{ht}
\alias{ht}
\title{Show first and last rows}
\usage{
ht(d, myrows = 10)
}
\arguments{
\item{d}{object to display}

\item{myrows}{how many of the first and last rows to show}
}
\value{
Matrix of the first and last myrow rows
}
\description{
Displays the first and last rows of a table
}
