% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNormTransform.R
\name{plotNormTransform}
\alias{plotNormTransform}
\title{plotNormTransform}
\usage{
plotNormTransform(eset_before, eset_after, titeldet = "", png_fn = NULL)
}
\arguments{
\item{eset_before}{TODO Erklaerung}

\item{eset_after}{TODO Erklaerung}

\item{png_fn}{TODO Erklaerung}
}
\value{
a png plot
}
\description{
This function plots an eset  befor and another eset after normalisation
}
