% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeBatchEffects.R
\name{removeBatchEffects}
\alias{removeBatchEffects}
\title{Remove Batch effects from expressoin data of a HT12prepro object}
\usage{
removeBatchEffects(
  ht12object,
  paramfile = NULL,
  excludeERCC = "from_paramfile",
  save_subgroupcontrast = "from_paramfile",
  second_combat_withvar = "from_paramfile",
  additional_normalisation_after_combat = "from_paramfile",
  normalisation_method = "from_paramfile",
  showPlot_additional_normalisation_after_combat = T
)
}
\arguments{
\item{ht12object}{A list object of class HT12prepro created with function filter4MinBatchsize}

\item{paramfile}{Path to the file specifying parameters}

\item{excludeERCC}{exclude ERCC samples from following steps, this is suggested if some samples have spiked-in ERCC samples but others have not. If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}

\item{save_subgroupcontrast}{Protect the subgroup contrast (column subgroup in the table stored in the slot $chipsamples of the HT12prepro object)from beeing affected in batch ajdustment, CAVE risk of false positives when subgroups are unbalanced with batches, hence FALSE is in this case recommended. If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}

\item{second_combat_withvar}{A second round of removeBatchEffects() can be intended. In this case, batches from the first round checking Sentrix.Barcode batches as well as batches from this second round specified as a column in ht12object$chipsamples named according to this variable are removed sequentially. If "", no 2nd combat will be done. If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}

\item{additional_normalisation_after_combat}{Indicator (TRUE or FALSE), whether to normalize data after batch-correction with Combat using the normalization method specified in parameter 'normalisation_method'. If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}

\item{normalisation_method}{Method used for normalisation if parameter additional_normalisation_after_combat = TRUE. Either 'quantile' (quantile normalisation) or 'rsn' (robust spline normalisation). If "from_paramfile", than the parameter will be read from the paramfile with the location of this file given in parameter paramfile.}
}
\value{
A list object of class HT12prepro where a slot with an expression set with normalized and transformed and batch-corrected data including control probe information is created. This slot is named `$total_nobkgd_eset_ql_combat`. Additionally, a graph is shown showing probe-wise as well as sample-wise correlation of expression data before and after batch adjustment.
}
\description{
This functions removes via ComBat from packge sva batch effects origniating from the Chip (Sentrix.Barcode) of the sample.  If second_combat_withvar is provided, a second round of removeBatchEffects() is done.
}
